DERBY-6117 Silently swallowed SecurityExceptions may disable Derby features, including security features.
Backported from trunk as: merge -c 1617737 https://svn.apache.org/repos/asf/db/derby/code/trunk

Patch derby-6617-3b. It adds reporting of the swallowed
SecurityException in FileMonitor#createDaemonGroup.

To get that printed I also i had to to make sure to dump the temporary
log in a change added to BaseMonitor#runWithState.

Additionally, I discovered that Derby became unbootable if we lack the
modifyThreadGroup permission: the monitor in such an event, lacking a
proper handler, thought it was already initialized so subsequent boot
attempts (from a non-system thread so we wouldn't see the
modifyThreadGroup issue) would also fail. I added a handler to clear
the monitor to fix this.

A new test fixture, testModifyThreadGroup checks that the warning
message being written to the console.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.11@1617738 13f79535-47bb-0310-9956-ffa450edef68